1. Change CADFEvent class to PyCadfEventMarshal 2. Remove new_model_tests.py and merge the old test into core/tests.py 3. Add additional test coverage for all the new functions in SavedSearch and DailyIndexDocType 4. Modify older tests to test for new functionality
