+++ goldstone/core/tests.py
@@ -19,21 +19,21 @@ This module demonstrates no less than 3 strategies for mocking ES.
 from django.conf import settings
 from django.test import SimpleTestCase, TestCase
 import elasticsearch
-from pycadf import event, cadftaxonomy
-from elasticsearch_dsl import Search
+import elasticsearch_dsl
+from elasticsearch_dsl import Search, String, Date, Nested
+from elasticsearch_dsl.result import Response
 import mock
 from mock import patch, MagicMock
-from rest_framework.test import APISimpleTestCase
-from goldstone.drfes.models import DailyIndexDocType
-
+from rest_framework.test import APISimpleTestCase, APITestCase
 from goldstone.test_utils import Setup
 from .models import Image, ServerGroup, NovaLimits, PolyResource, Host, \
-    Aggregate, Hypervisor, Port, Cloudpipe, Network, Project, Server, Addon, \
-    SavedSearch, EventQueryDef, PycadfEventConverter
-from goldstone.tests import ElasticSearchTests
-
+    Aggregate, Hypervisor, Port, Cloudpipe, Network, Project, Server, Addon
 from . import tasks
 from .utils import custom_exception_handler, process_resource_type, parse
+from goldstone.drfes.new_models import DailyIndexDocType
+from goldstone.core.new_models import SavedSearch, PyCadfEventMarshal
+from pycadf import event, cadftype, cadftaxonomy
+import uuid
 
 # Using the latest version of django-polymorphic, a
 # PolyResource.objects.all().delete() throws an IntegrityError exception. So
@@ -98,43 +98,207 @@ class TaskTests(SimpleTestCase):
         self.assertEqual(tasks.delete_indices('abc', 10), 'mocked')
 
 
-class PycadfEventConverterTests(SimpleTestCase):
+class DailyIndexDocTypeTests(SimpleTestCase):
+
+    class CADFEvent(DailyIndexDocType):
+
+        INDEX_DATE_FMT = 'YYYY-MM-DD'
+
+        traits = Nested(
+            properties={
+                'action': String(),
+                'eventTime': Date(),
+                'eventType': String(),
+                'id': String(),
+                'initiatorId': String(),
+                'name': String(),
+                'observerId': String(),
+                'outcome': String(),
+                'targetId': String(),
+                'typeURI': String(),
+            }
+        )
+
+        class Meta:
+            doc_type = 'cadf_event'
+            index = 'events_*'
+
+        def __init__(self, event=None, meta=None, **kwargs):
+            if event is not None:
+                kwargs = dict(
+                    kwargs.items() + self._get_traits_dict(event).items())
+
+            super(DailyIndexDocTypeTests.CADFEvent, self) \
+                .__init__(meta, **kwargs)
+
+        @staticmethod
+        def _get_traits_dict(e):
+            """
+            convert a pycadf.event to an ES doc
+            :param e:
+            :return: dict
+            """
+            return {"traits": e.as_dict()}
+
+    class LogMessage(DailyIndexDocType):
+
+        message = String()
+        timestamp = Date()
+
+        class Meta:
+            doc_type = 'syslog'
+            index = 'logstash-*'
+
+    def test_search_wildcard_index(self):
+        s = self.LogMessage.search()
+        result = s.execute()
+        # verify that execute() returns an es_dsl result object
+        self.assertIsInstance(result, Response)
+
+    def test_index_today(self):
+        import arrow
+        date_fmt = self.LogMessage.INDEX_DATE_FMT
+        today = arrow.utcnow().format(date_fmt)
+        dt = self.LogMessage()
+        self.assertEqual(dt._index_today(), 'logstash-' + today)
+
+    def test_lifecycle(self):
+        import arrow
+
+        # to avoid worrying about searching the raw field, let's use a
+        # lowercase string with no special chars.
+        message = ''.join([c for c in str(uuid.uuid4()) if c != '-'])
+        print "test_lifecycle() : message "
+        print message
+        created = arrow.utcnow().isoformat()
+
+        # instantiate and persist the LogMessage
+        lm = self.LogMessage(message=message, created=created)
+        result = lm.save()
+        # verify that calling save() on es_dsl results in persistence
+        self.assertEqual(result, True)
+
+        # force flush the index so our test has a chance to succeed.
+        lm._doc_type.using.indices.flush(lm.meta.index)
+
+        # Sadly, the get method does not accept a wildcard index, so you will
+        # need to either know which index the document is in, or use the search
+        # method to get it back.  When you save an instance, the meta attrs
+        # id and index are set, so if it's important information, you may
+        # want to persist it in postgres.
+        glm = self.LogMessage.get(id=lm.meta.id, index=lm.meta.index)
+        # Method 1 : verify that what we obtain by calling a get() with
+        # index-filtering is indeed the log-message we persisted earlier
+        self.assertEqual(lm.message, glm.message)
+        self.assertEqual(lm.created, glm.created)
+        self.assertIsInstance(glm, self.LogMessage)
+
+        # so let's try to find it via search.
+        s = self.LogMessage.search()
+        # filter by the string log-message we used for creating a LogMessage
+        # instance. This object was used earlier to create and persist a new
+        # log-message in DailyIndexDocType.
+        slm = s.filter('term', message=message) \
+            .filter('term', created=created) \
+            .execute()[0]
+
+        # Method 2 : verify that what we got by searching and filtering
+        # by message is indeed the log-message we persisted earlier
+        self.assertEqual(lm.message, slm.message)
+        self.assertEqual(lm.created, slm.created)
+        self.assertIsInstance(slm, self.LogMessage)
+
+        # let's make sure we can delete an object
+        result = slm.delete()
+        self.assertIsNone(result)
+
+        # force flush the index so our test has a chance to succeed.
+        lm._doc_type.using.indices.flush(lm.meta.index)
+
+        # we should not be able to find the record now.
+        s = self.LogMessage.search()
+        slm2 = s.filter('term', message=message) \
+            .filter('term', created=created) \
+            .execute()
+        self.assertEqual(len(slm2.hits), 0)
+
+    def test_cadf_event(self):
+        initiator_id = ''.join([c for c in str(uuid.uuid4()) if c != '-'])
+        observer_id = ''.join([c for c in str(uuid.uuid4()) if c != '-'])
+        target_id = ''.join([c for c in str(uuid.uuid4()) if c != '-'])
+
+        # create a CADF event something like (using .as_dict()):
+        # {'action': 'read',
+        #  'eventTime': '2015-12-21T18:47:50.275715+0000',
+        #  'eventType': 'activity',
+        #  'id': '2f38134e-c880-5a38-8b3e-101554a71e37',
+        #  'initiatorId': '46e4be801e0b4c2ab373b26dceedce1a',
+        #  'name': 'test event',
+        #  'observerId': 'a6b012069d174d4fbf9acee03367f068',
+        #  'outcome': 'success',
+        #  'targetId': '005343040e084c3ba90f7bac1b97e1ae',
+        #  'typeURI': 'http://schemas.dmtf.org/cloud/audit/1.0/event'}
+        e = event.Event(
+            eventType=cadftype.EVENTTYPE_ACTIVITY,
+            action=cadftaxonomy.ACTION_READ,
+            outcome=cadftaxonomy.OUTCOME_SUCCESS,
+            initiatorId=initiator_id,
+            targetId=target_id,
+            observerId=observer_id,
+            name="test event")
+
+        cadf = self.CADFEvent(event=e)
+
+        result = cadf.save()
+        self.assertTrue(result)
+
+        # force flush the index so our test has a chance to succeed.
+        cadf._doc_type.using.indices.flush(cadf.meta.index)
+
+        ge = self.CADFEvent.get(id=cadf.meta.id, index=cadf.meta.index)
+        self.assertEqual(cadf.traits['initiatorId'], ge.traits['initiatorId'])
+        self.assertEqual(cadf.traits['eventTime'], ge.traits['eventTime'])
+        self.assertIsInstance(ge, self.CADFEvent)
+
+        s = self.CADFEvent.search()
+        s = s.filter('term', ** {'traits.initiatorId': initiator_id})\
+            .execute()
+        self.assertIsInstance(s, Response)
+        self.assertEqual(len(s.hits), 1)
+
+        # let's make sure we can delete an object
+        result = cadf.delete()
+        self.assertIsNone(result)
+
+        # force flush the index so our test has a chance to succeed.
+        cadf._doc_type.using.indices.flush(cadf.meta.index)
+
+        # we should not be able to find the record now.
+        s2 = self.CADFEvent.search()
+        # Alternate form of filter/query expression since we have a nested
+        # field.
+        r2 = s2.filter('term', ** {'traits.initiatorId': initiator_id})\
+            .execute()
+        self.assertEqual(len(r2.hits), 0)
+
+
+class PyCadfEventMarshalTests(SimpleTestCase):
     """
-        Test PycadfEventConverter class for correctness
+        Test PyCadfEventMarshal class for correctness
     """
 
     @classmethod
     def test_event_save(cls):
-        converter = PycadfEventConverter()
-        event_saved_flag = converter.save(converter)
-        print event_saved_flag
-        assert(event_saved_flag)
-
-
-class PycadflibraryTests(SimpleTestCase):
-    """ Test pycadf library for correctness
-
-    """
-    pycadf_event = ''
-
-    @classmethod
-    def test_pycadf_event_create(self):
-
+        pyobj = PyCadfEventMarshal()
+        isinstance(pyobj, PyCadfEventMarshal)
         new_event = event.Event()
-        print new_event
-        assert(new_event)
         isinstance(new_event, event.Event)
-        assert(new_event.id)
-        self.pycadf_event = new_event
-
-    @classmethod
-    def test_pycadf_event_save(self):
-        d = DailyIndexDocType()
-        save_flag = True
-        # TBD : To test save() es_dsl connection needs to be established.
-        # save_flag = d.save(index='events_2015-12-18',
-        #                   event_instance=self.pycadf_event)
-        assert(save_flag)
+        trait_dict = pyobj._get_traits_dict(new_event)
+        isinstance(trait_dict, dict)
+        assert('id' in trait_dict['traits'])
+        assert('eventTime' in trait_dict['traits'])
+        event_saved_flag = pyobj.save(pyobj)
+        assert(event_saved_flag)
 
 
 class SavedSearchModelTests(TestCase):
@@ -150,29 +314,36 @@ class SavedSearchModelTests(TestCase):
 
     fixtures = ['initial_data.yaml']
 
-    def setUp(self):
-        setup_flag = True
-        assert(setup_flag)
+    message = String()
+    timestamp = Date()
+
+    class Meta:
+        doc_type = 'syslog'
+        index = 'logstash-*'
 
     @classmethod
     def test_loaded_data_from_fixtures(cls):
         assert(SavedSearch.objects.all())
 
     @classmethod
-    def test_execute_query_func(self):
-        savedsearch_instance = SavedSearch(name='test_savedsearch')
-        rv = savedsearch_instance.execute_query('sys_log_err_type_search_1')
-        assert isinstance(rv, Search)
-        assert (rv)
-        # TBD : Compare expected vs returned results
-        # self.assertIsNotNone(rv,
-        # msg="Elastic Search query returned null result")
+    def test_predefined_search_func(self):
+        owner = 'events'
+        indices = 'logstash-*'
+        sys_defined_searches = SavedSearch.objects.filter(owner=owner,
+                                                          index_prefix=indices)
+        assert(sys_defined_searches)
+        isinstance(sys_defined_searches, SavedSearch)
+        for entry in sys_defined_searches:
+            search_obj = entry.search()
+            isinstance(search_obj, elasticsearch_dsl.search.Search)
+            search_res = search_obj.execute()
+            assert(search_res.hits.total)
 
-    @classmethod
-    def test_search_by_syslog_err_type_func(cls):
-        event_query = EventQueryDef()
-        rv = event_query.search_by_syslog_err_type()
-        assert (rv)
+        for entry in sys_defined_searches:
+            search_obj, start, end = entry.search_recent()
+            isinstance(search_obj, elasticsearch_dsl.search.Search)
+            search_res = search_obj.execute()
+            assert(search_res.hits.total)
 
 
 class PolyResourceModelTests(SimpleTestCase):
